elt/src/transform/requirements.txt
azure_storage==0.37.0
numpy==1.21.5
pandas==1.4.4
pyodbc==4.0.34
python-dotenv==1.0.0
SQLAlchemy==1.4.39
